# specialcase for bfloat16->float16 cast