--- a/tinygrad/tensor.py
+++ b/tinygrad/tensor.py
@@ -3727,8 +3727,7 @@ class Tensor(SimpleMathTrait):
       # NOTE: values within the int32 range and outside the unsigned dtype range will cause values to wrap around
       return self._apply_uop(UOp.cast, dtype=dtypes.int32)._apply_uop(UOp.cast, dtype=dt)
     if self.dtype == dtypes.bfloat16 and dt == dtypes.float16:
-        return (self._apply_uop(UOp.cast, dtype=dtypes.float32)
-                    ._apply_uop(UOp.cast, dtype=dtypes.float16))
+        return (self._apply_uop(UOp.cast, dtype=dtypes.float32)._apply_uop(UOp.cast, dtype=dtypes.float16))
     return self if self.dtype == dt else self._apply_uop(UOp.cast, dtype=dt)
 
   def bitcast(self, dtype:DTypeLike) -> Tensor: